+++ mapie/regression/time_series_regression.py
@@ -1,13 +1,12 @@
 from __future__ import annotations
 
-from typing import Iterable, Optional, Tuple, Union, cast
+from typing import Optional, Union, cast
 
 import numpy as np
 from sklearn.base import RegressorMixin
 from sklearn.model_selection import BaseCrossValidator
 from sklearn.utils.validation import check_is_fitted
 
-from mapie._compatibility import np_nanquantile
 from mapie._typing import ArrayLike, NDArray
 from mapie.aggregation_functions import aggregate_all
 from mapie.conformity_scores import ConformityScore
@@ -62,6 +61,7 @@ class MapieTimeSeriesRegressor(MapieRegressor):
 
     cv_need_agg_function_ = MapieRegressor.cv_need_agg_function_ + ["BlockBootstrap"]
     valid_methods_ = ["enbpi", "aci"]
+    default_sym_ = False
 
     def __init__(
         self,
@@ -89,90 +89,40 @@ class MapieTimeSeriesRegressor(MapieRegressor):
         self,
         X: ArrayLike,
         y: ArrayLike,
+        ensemble: bool = False,
     ) -> NDArray:
         """
         Compute the conformity scores on a data set.
 
         Parameters
         ----------
-            X : ArrayLike of shape (n_samples, n_features)
+        X : ArrayLike of shape (n_samples, n_features)
             Input data.
 
-            y : ArrayLike of shape (n_samples,)
+        y : ArrayLike of shape (n_samples,)
                 Input labels.
+
+        ensemble: bool
+            Boolean determining whether the predictions are ensembled or not.
+            If ``False``, predictions are those of the model trained on the
+            whole training set.
+            If ``True``, predictions from perturbed models are aggregated by
+            the aggregation function specified in the ``agg_function``
+            attribute.
+
+            If ``cv`` is ``"prefit"`` or ``"split"``, ``ensemble`` is ignored.
+
+            By default ``False``.
 
         Returns
         -------
             The conformity scores corresponding to the input data set.
         """
-        y_pred = super().predict(X, ensemble=True)
+        y_pred = super().predict(X, ensemble=ensemble)
         scores = np.array(
-            self.conformity_score_function_.get_signed_conformity_scores(X, y, y_pred)
+            self.conformity_score_function_.get_conformity_scores(X, y, y_pred)
         )
         return scores
-
-    def _beta_optimize(
-        self,
-        alpha: Union[float, NDArray],
-        upper_bounds: NDArray,
-        lower_bounds: NDArray,
-    ) -> NDArray:
-        """
-        Minimize the width of the PIs, for a given difference of quantiles.
-
-        Parameters
-        ----------
-        alpha: Union[float, NDArray]
-            The quantiles to compute.
-        upper_bounds: NDArray
-            The array of upper values.
-        lower_bounds: NDArray
-            The array of lower values.
-
-        Returns
-        -------
-        NDArray
-            Array of betas minimizing the differences
-            ``(1-alpa+beta)-quantile - beta-quantile``.
-
-        Raises
-        ------
-        ValueError
-            If lower and upper bounds arrays don't have the same shape.
-        """
-        if lower_bounds.shape != upper_bounds.shape:
-            raise ValueError(
-                "Lower and upper bounds arrays should have the same shape."
-            )
-        alpha = cast(NDArray, alpha)
-        betas_0 = np.full(
-            shape=(len(lower_bounds), len(alpha)),
-            fill_value=np.nan,
-            dtype=float,
-        )
-
-        for ind_alpha, _alpha in enumerate(alpha):
-            betas = np.linspace(
-                _alpha / (len(lower_bounds) + 1),
-                _alpha,
-                num=len(lower_bounds),
-                endpoint=True,
-            )
-            one_alpha_beta = np_nanquantile(
-                upper_bounds.astype(float),
-                1 - _alpha + betas,
-                axis=1,
-                method="higher",
-            )
-            beta = np_nanquantile(
-                lower_bounds.astype(float),
-                betas,
-                axis=1,
-                method="lower",
-            )
-            betas_0[:, ind_alpha] = betas[np.argmin(one_alpha_beta - beta, axis=0)]
-
-        return betas_0
 
     def init_alpha(self) -> None:
         if self.method == "aci":
@@ -228,6 +178,7 @@ class MapieTimeSeriesRegressor(MapieRegressor):
         X: ArrayLike,
         y: ArrayLike,
         sample_weight: Optional[ArrayLike] = None,
+        ensemble: bool = False,
     ) -> MapieTimeSeriesRegressor:
         """
         Compared to the method ``fit`` of ``MapieRegressor``, the ``fit``
@@ -253,13 +204,27 @@ class MapieTimeSeriesRegressor(MapieRegressor):
 
             By default ``None``.
 
+        ensemble: bool
+            Boolean determining whether the predictions are ensembled or not.
+            If ``False``, predictions are those of the model trained on the
+            whole training set.
+            If ``True``, predictions from perturbed models are aggregated by
+            the aggregation function specified in the ``agg_function``
+            attribute.
+
+            If ``cv`` is ``"prefit"`` or ``"split"``, ``ensemble`` is ignored.
+
+            By default ``False``.
+
         Returns
         -------
         MapieTimeSeriesRegressor
             The model itself.
         """
         self = super().fit(X=X, y=y, sample_weight=sample_weight)
-        self.conformity_scores_ = self._relative_conformity_scores(X, y)
+        self.conformity_scores_ = self._relative_conformity_scores(
+            X, y, ensemble=ensemble
+        )
         if self.method == "aci":
             self.current_alpha: dict[float, float] = {}
         return self
@@ -268,6 +233,7 @@ class MapieTimeSeriesRegressor(MapieRegressor):
         self,
         X: ArrayLike,
         y: ArrayLike,
+        ensemble: bool = False,
     ) -> MapieTimeSeriesRegressor:
         """
         Update the ``conformity_scores_`` attribute when new data with known
@@ -281,6 +247,18 @@ class MapieTimeSeriesRegressor(MapieRegressor):
 
         y: ArrayLike of shape (n_samples_test,)
             Input labels.
+
+        ensemble: bool
+            Boolean determining whether the predictions are ensembled or not.
+            If ``False``, predictions are those of the model trained on the
+            whole training set.
+            If ``True``, predictions from perturbed models are aggregated by
+            the aggregation function specified in the ``agg_function``
+            attribute.
+
+            If ``cv`` is ``"prefit"`` or ``"split"``, ``ensemble`` is ignored.
+
+            By default ``False``.
 
         Returns
         -------
@@ -302,7 +280,9 @@ class MapieTimeSeriesRegressor(MapieRegressor):
                 "The number of observations to update is higher than the"
                 "number of training instances."
             )
-        new_conformity_scores_ = self._relative_conformity_scores(X, y)
+        new_conformity_scores_ = self._relative_conformity_scores(
+            X, y, ensemble=ensemble
+        )
         self.conformity_scores_ = np.roll(
             self.conformity_scores_, -len(new_conformity_scores_)
         )
@@ -378,121 +358,6 @@ class MapieTimeSeriesRegressor(MapieRegressor):
                 self.current_alpha[alpha_0] = new_alpha_t
 
         return self
-
-    def predict(
-        self,
-        X: ArrayLike,
-        ensemble: bool = False,
-        alpha: Optional[Union[float, Iterable[float]]] = None,
-        optimize_beta: bool = True,
-    ) -> Union[NDArray, Tuple[NDArray, NDArray]]:
-        """
-        Correspond to 'Conformal prediction for dynamic time-series'.
-
-        Parameters
-        ----------
-        X: ArrayLike of shape (n_samples, n_features)
-            Test data.
-
-        ensemble: bool
-            Boolean determining whether the predictions are ensembled or not.
-            If ``False``, predictions are those of the model trained on the
-            whole training set.
-            If ``True``, predictions from perturbed models are aggregated by
-            the aggregation function specified in the ``agg_function``
-            attribute.
-
-            If ``cv`` is ``"prefit"`` or ``"split"``, ``ensemble`` is ignored.
-
-            By default ``False``.
-
-        alpha: Optional[Union[float, Iterable[float]]]
-            Can be a float, a list of floats, or a ``ArrayLike`` of floats.
-            Between ``0`` and ``1``, represents the uncertainty of the
-            confidence interval.
-            Lower ``alpha`` produce larger (more conservative) prediction
-            intervals.
-            ``alpha`` is the complement of the target coverage level.
-
-            By default ``None``.
-
-        optimize_beta: bool
-            Whether to optimize the PIs' width or not.
-
-        Returns
-        -------
-        Union[NDArray, Tuple[NDArray, NDArray]]
-            - NDArray of shape (n_samples,) if ``alpha`` is ``None``.
-            - Tuple[NDArray, NDArray] of shapes (n_samples,) and
-              (n_samples, 2, n_alpha) if ``alpha`` is not ``None``.
-                - [:, 0, :]: Lower bound of the prediction interval.
-                - [:, 1, :]: Upper bound of the prediction interval.
-        """
-        # Checks
-        check_is_fitted(self, self.fit_attributes)
-        self._check_ensemble(ensemble)
-        alpha = cast(Optional[NDArray], check_alpha(alpha))
-        y_pred = self.estimator_.single_estimator_.predict(X)
-        n = len(self.conformity_scores_)
-
-        if alpha is None:
-            return np.array(y_pred)
-
-        alpha_np = cast(NDArray, alpha)
-        check_alpha_and_n_samples(alpha_np, n)
-        alpha_np = np.round(alpha_np, 2)
-
-        if self.method == "aci":
-            for ix, alpha_checked in enumerate(alpha_np):
-                # This code snippet in the "aci" method ensures that when
-                # the same confidence level (alpha) is encountered more
-                # than once, it is mapped to a consistent value.
-                # This helps maintain reliability and predictability in
-                # the algorithm's computations specific to the "aci" method.
-                if alpha_checked not in self.current_alpha:
-                    self.current_alpha[alpha_checked] = alpha_checked
-                else:
-                    alpha_np[ix] = self.current_alpha[alpha_checked]
-
-        if optimize_beta:
-            betas_0 = self._beta_optimize(
-                alpha_np,
-                self.conformity_scores_.reshape(1, -1),
-                self.conformity_scores_.reshape(1, -1),
-            )
-        else:
-            betas_0 = np.repeat(alpha[:, np.newaxis] / 2, n, axis=0)
-
-        lower_quantiles = np_nanquantile(
-            self.conformity_scores_.astype(float),
-            betas_0[0, :],
-            axis=0,
-            method="lower",
-        ).T
-        higher_quantiles = np_nanquantile(
-            self.conformity_scores_.astype(float),
-            1 - alpha_np + betas_0[0, :],
-            axis=0,
-            method="higher",
-        ).T
-        self.lower_quantiles_ = lower_quantiles
-        self.higher_quantiles_ = higher_quantiles
-
-        if self.method in self.no_agg_methods_ or self.cv in self.no_agg_cv_:
-            y_pred_low = y_pred[:, np.newaxis] + lower_quantiles
-            y_pred_up = y_pred[:, np.newaxis] + higher_quantiles
-        else:
-            y_pred_multi = self.estimator_._pred_multi(X)
-            pred = aggregate_all(self.agg_function, y_pred_multi)
-            lower_bounds, upper_bounds = pred, pred
-
-            y_pred_low = lower_bounds.reshape(-1, 1) + lower_quantiles
-            y_pred_up = upper_bounds.reshape(-1, 1) + higher_quantiles
-
-            if ensemble:
-                y_pred = aggregate_all(self.agg_function, y_pred_multi)
-
-        return y_pred, np.stack([y_pred_low, y_pred_up], axis=1)
 
     def _more_tags(self):
         return {